Wrote two versions, settled on one:
Initial implementation would take the JSON files, convert them to
Sereal binary format, then read all of it into memory. There is
an import script for it.

This is suboptimal, at best, because it was a lot of data to load
all into memory. I would also have to implement a query mechanism
like a DB.

Second implementation was to just use a DB. There is an import
script for that too. It's *much* slower than using the Sereal files.
I'm not sure why but I might be using SQLite incorrectly.

I rewrote the modules so the entire Sereal support is gone, but
the import script is good enough to backtrack if you need to.

Next up:
* More tests.
* Documentation.
* Import everything and commit SQLite DB too.
* Configure share directory so it can be released to CPAN.

But, I mean, it's a good start. Especially for something I did not
intend to do today. :)